requirements.txt
fastapi
uvicorn
pydantic
motor
pyjwt
python-dotenv
passlib[bcrypt]